requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
discord==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
numpy==1.26.4
pandas==2.2.2
py-cord==2.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
yarl==1.9.4